+++ R/np.density.R
@@ -315,7 +315,7 @@ npudens.default <- function(bws, tdat, ..., se = FALSE){
       !tdat.formula.early && !inherits(formula.input, "formula"))
     return(.npRmpi_autodispatch_call(match.call(), parent.frame()))
 
-  sc <- sys.call()
+  sc <- .np_formula_expand_call(sys.call(), parent.frame())
   sc.names <- names(sc)
 
   ## here we check to see if the function was called with tdat =
@@ -372,10 +372,12 @@ npudens.default <- function(bws, tdat, ..., se = FALSE){
   tbw <- if (!has.explicit.bws) {
     .np_progress_select_bandwidth_enhanced(
       "Selecting density bandwidth",
-      .np_eval_bw_call(sc.bw, caller_env = parent.frame())
+      .np_eval_bw_call(sc.bw, caller_env = parent.frame(),
+                         formula.value = .np_formula_value(formula.input, bws, tdat, "dat"))
     )
   } else {
-    .np_eval_bw_call(sc.bw, caller_env = parent.frame())
+    .np_eval_bw_call(sc.bw, caller_env = parent.frame(),
+                         formula.value = .np_formula_value(formula.input, bws, tdat, "dat"))
   }
 
   ## convention: first argument is always dropped, second, if present, propagated