+++ R/np.density.bw.R
@@ -5,7 +5,8 @@ npudensbw <- function(...){
   .np_validate_public_dots(mc[["..."]], "npudensbw")
   target <- .np_bw_dispatch_target(dots = mc$...,
                                    data_arg_names = "dat",
-                                   eval_env = parent.frame())
+                                     eval_env = parent.frame(),
+                                     promise.frame = environment())
   UseMethod("npudensbw", target)
 }
 
@@ -37,6 +38,7 @@ npudensbw.formula <-
 
     mf[[1]] <- as.name("model.frame")
     mf.args <- as.list(mf[-1L])
+    mf.args[["formula"]] <- substitute(quote(VALUE), list(VALUE = formula))
     mf <- do.call(.np_formula_model_frame, mf.args, envir = parent.frame())
 
     if (attr(attr(mf, "terms"), "response") != 0)