+++ R/np.distribution.bw.R
@@ -5,7 +5,8 @@ npudistbw <- function(...){
   .np_validate_public_dots(mc[["..."]], "npudistbw")
   target <- .np_bw_dispatch_target(dots = mc$...,
                                    data_arg_names = c("dat", "gdat"),
-                                   eval_env = parent.frame())
+                                     eval_env = parent.frame(),
+                                     promise.frame = environment())
   UseMethod("npudistbw", target)
 }
 
@@ -79,6 +80,7 @@ npudistbw.formula <-
 
     mf[[1]] <- as.name("model.frame")
     mf.args <- as.list(mf[-1L])
+    mf.args[["formula"]] <- substitute(quote(VALUE), list(VALUE = formula))
     mf <- do.call(.np_formula_model_frame, mf.args, envir = parent.frame())
 
     if (attr(attr(mf, "terms"), "response") != 0)